pip-requirements.txt
git+https://github.com/guluc3m/zoe-lib-python.git
ansicolors
